ac_1.1
This version was tested and working right after the fishing update.

HOW TO USE
The bot allows you to auto-type and auto-catch specific Pokemon as long as you have Discord open and the chat bar is selected. To begin, first add the bot to your Discord server and ensure that this program is running. To begin, send "start" to the server with your bot and do not click out of the chat bar on Discord. To pause, type "start" and to stop, type "stop".
Make sure to insert your own Bot token, Discord id, and Discord Tag.

HOW IT WORKS
The auto-catching works by determining what colour code is used for the embed message and depending on this, the bot throws different types of balls. The current setup is:
Greatball for Common, Uncommon, and Rare (if streak is not one less than a multiple of 5)
Ultraball for Rare (if streak is one less than a multiple of 5), Super-rare, Legendary (that is already caught)
Premier-ball for Full-odds Shiny
Masterball for Event Shiny, Legendary (that hasn't been caught), and special Pokemon (Arceus, Shaymin, Darkrai, etc.)

The bot also auto-buys Greatballs and Ultraballs when their quantities reach below 50 and 10 respectively.

The fishing aspect of the bot was only tested for a few days after the update released, so this might be buggy. The bot auto-fishes every 2 spawns of regular Pokemon and also auto-pull the rod when a Pokemon has been caught. The shiny Pokemon detection in fishing was not properly tested (since shinies from fishing is extremely rare), so this might be buggy as well.

NOTIFICATIONS
The user is notified through Discord DM whenever they encounter a Legendary, Shiny, or Special Pokemon, and when an egg is ready to hatch. The user is also notified when a captcha appears.